Add sparse jacobian templates to AnyODE
This seems to be the least invasive/most generic way of adding
sparse (CSR or CSC) matrices to AnyODE. All one needs to specify
is i) the type used for the index values and ii) the max number of
nonzero entries in the jacobian (nnz), which is regarded as a
constant like ny. Conditional on these pieces of information,
a fairly generic signature for a CSR/CSC matrix is as shown.